requirements.txt
python-dotenv
fastapi[all]
uvicorn
pipecat-ai[google,silero,webrtc]>=0.0.82